codes

+++ Springboot framework/wallet/src/main/java/com/transmoney/wallet/controller/AccountController.java
@@ -38,6 +38,9 @@ public class AccountController {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<String> delete(@PathVariable Long id){
+        if(!service.accountExist(id))
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found.");
+        
         service.deleteAccount(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
